Move displayhook for ZMQ shell to zmq.displayhook, and rename to make the difference clearer.
